fix: renaming files and imports, fix dependabot config
