+++ package.json
@@ -67,19 +67,19 @@
     "@rollup/plugin-typescript": "^12.1.2",
     "@storyblok/eslint-config": "^0.3.0",
     "@sveltejs/adapter-auto": "^3.3.1",
-    "@sveltejs/kit": "^2.15.2",
+    "@sveltejs/kit": "^2.15.3",
     "@sveltejs/package": "^2.3.7",
     "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "cypress": "^13.17.0",
     "eslint": "^9.18.0",
-    "eslint-plugin-svelte": "^2.36.0",
+    "eslint-plugin-svelte": "^2.46.1",
     "jsdom": "^26.0.0",
-    "publint": "^0.2.0",
+    "publint": "^0.3.2",
     "simple-git-hooks": "^2.11.1",
     "start-server-and-test": "^2.0.10",
-    "svelte": "^5.17.4",
+    "svelte": "^5.18.0",
     "typescript": "^5.7.3",
-    "typescript-eslint": "^8.19.1",
+    "typescript-eslint": "^8.20.0",
     "vite": "^6.0.7",
     "vite-plugin-dts": "^4.5.0",
     "vitest": "^2.1.8"
